+++ configs/datasets/gpqa/gpqa_gen.py
@@ -1,4 +1,4 @@
 from mmengine.config import read_base
 
 with read_base():
-    from .gpqa_gen_5aeece import gpqa_datasets
+    from .gpqa_openai_simple_evals_gen_5aeece import gpqa_datasets
